% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_settings_server.R
\name{capture_python_output}
\alias{capture_python_output}
\title{Capture python output}
\usage{
capture_python_output(code)
}
\arguments{
\item{code}{Python code that you want to run (character)}
}
\description{
A function to get Python console result
}
