% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_layout_functions.R
\name{create_hover_card}
\alias{create_hover_card}
\title{Create a hover card}
\usage{
create_hover_card(ui = character(), text = character())
}
\arguments{
\item{ui}{UI element (character)}

\item{text}{Text to display in the card (character)}
}
\description{
Displays a card when hovering an UI element
}
