% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_model.R
\name{get_omop_col_names}
\alias{get_omop_col_names}
\title{Get OMOP column names per table for a given version}
\usage{
get_omop_col_names(version = "5.4")
}
\arguments{
\item{version}{OMOP CDM version ("5.3" or "5.4")}
}
\value{
Named list of character vectors (column names)
}
\description{
Get OMOP column names per table for a given version
}
