% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_model.R
\name{get_omop_col_types}
\alias{get_omop_col_types}
\title{Get OMOP column types per table for a given version}
\usage{
get_omop_col_types(version = "5.4")
}
\arguments{
\item{version}{OMOP CDM version ("5.3" or "5.4")}
}
\value{
Named list of character strings encoding column types
}
\description{
Get OMOP column types per table for a given version
}
