% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_model.R
\name{get_omop_vocabulary_tables}
\alias{get_omop_vocabulary_tables}
\title{Get OMOP vocabulary table names}
\usage{
get_omop_vocabulary_tables()
}
\value{
A character vector of OMOP vocabulary table names
}
\description{
Get OMOP vocabulary table names
}
