% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{get_query}
\alias{get_query}
\title{Query OMOP data from a database connection}
\usage{
get_query(query)
}
\arguments{
\item{query}{A string containing the SQL query to execute}
}
\value{
A tibble containing the query results
}
\description{
This function executes SQL queries against an OMOP CDM database using the connection
stored in the global 'd' object (d$con).
}
\examples{
get_query("SELECT * FROM person")

}
