% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{is_integer_or_integer64}
\alias{is_integer_or_integer64}
\title{Is integer}
\usage{
is_integer_or_integer64(x)
}
\arguments{
\item{x}{Number to test (int / int64 / numeric)}
}
\description{
Test if a number is an integer or an integer64
}
\examples{
is_integer_or_integer64(634L)
}
