\name{linkR-package}
\alias{linkR-package}

\alias{analyzeLinkage}
\alias{angleOnCircleFromPoint}
\alias{avectors}
\alias{circlePoint}
\alias{combineLinkages}
\alias{copyTransformation}
\alias{cprod}
\alias{defineCircle}
\alias{distPointToLine}
\alias{distPointToPlane}
\alias{distPointToPoint}
\alias{drawLinkage}
\alias{findOptimalPointAlignment}
\alias{intersectCircles}
\alias{intersectSphereLine}
\alias{linkR_examples}
\alias{modelStaticForces}
\alias{pointNormalOnLine}
\alias{pointOnPlaneFromPoints}
\alias{pointPlaneProj}
\alias{proj3DTo2D}
\alias{reverseLinkage}
\alias{rotateBody}
\alias{rotationMatrixToEP}
\alias{rotationMatrixZYX}
\alias{setCoordinateAxes}
\alias{tMatrixDC}
\alias{tMatrixEP}
\alias{uvector}
\alias{vectorsToEP}
\alias{vorthogonal}

\docType{package}
\title{3D lever and linkage mechanism modeling}
\description{The linkR package for R provides a toolkit for creating models of 2D and 3D linkages, predicting motion, force and torque outputs and calculating linkage mechanical properties such as motion, force and torque transmission ratios. Users in the fields of engineering and biomechanics will find linkR particularly useful. linkR works with the R package \href{http://cran.r-project.org/web/packages/svgViewR/index.html}{svgViewR}, enabling users to create 3D interactive animations of linkage models.}
\details{
\tabular{ll}{
Package: \tab linkR\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2014-12-31\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Aaron Olsen
Maintainer: Aaron Olsen <aolsen@uchicago.edu>
}