% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkr.R
\name{linkr}
\alias{linkr}
\title{Run the Shiny application}
\usage{
linkr(
  language = "en",
  app_folder = character(),
  username = "admin",
  local = FALSE,
  debug = FALSE,
  log_file = FALSE,
  port = 3838,
  loading_options = list()
)
}
\arguments{
\item{language}{Default language to use in the App (character)}

\item{app_folder}{Location of the application folder (character).}

\item{username}{Username used for connection. Default username is "admin". (character)}

\item{local}{Run the app in local mode, do not load files on the internet (logical)}

\item{debug}{Debug mode : steps and errors will by displayed in the console (logical)}

\item{log_file}{Create a log file to see app log directly in the app (logical)}

\item{port}{Port used by shiny app (integer)}

\item{loading_options}{List specifying the initial page, project, and subset to load at startup (a list with `page`, `project`, and `subset` keys)}
}
\description{
Runs the LinkR Shiny Application.\cr
Use language argument to choose language to use ("en" or "fr" available).\cr
Use app_folder argument to choose a folder where the files of the application will be saved. By default, 
a 'linkr' folder will be created in the folder returned by path.expand("~").\cr
}
\examples{
\dontrun{
linkr(language = "en", debug = FALSE, local = FALSE)
}
}
