% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{now}
\alias{now}
\title{Get system time}
\usage{
now(format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{format}{Format to apply to Sys.time() result}
}
\description{
Get system time with Sys.time() function
}
\examples{
now()
}
