% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{resize_and_pad_image}
\alias{resize_and_pad_image}
\title{Resize and pad image}
\usage{
resize_and_pad_image(
  input_path,
  output_path,
  target_width = 318,
  target_height = 200,
  i18n = character()
)
}
\arguments{
\item{input_path}{Path of the input image}

\item{output_path}{Path of the output image}

\item{target_width}{Width of the output image}

\item{target_height}{Height of the output image}
}
\value{
Returns a PNG image with new dimensions
}
\description{
Resize and pad a PNG image
}
\examples{
resize_and_pad_image("my_image.png", "my_image_resized.png", target_width = 318, target_height = 200)
}
