% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_management.R
\name{add_patients_to_subset}
\alias{add_patients_to_subset}
\title{Add patients to a subset}
\usage{
add_patients_to_subset(
  output,
  r = shiny::reactiveValues(),
  m = shiny::reactiveValues(),
  patients = tibble::tibble(),
  subset_id = integer(),
  i18n = character(),
  ns = character()
)
}
\arguments{
\item{output}{Shiny output variable}

\item{r}{A shiny::reactiveValues object, used to communicate between modules}

\item{m}{A shiny::reactiveValues object, used to communicate between modules}

\item{patients}{numeric vector containing patients / persons (numeric)}

\item{subset_id}{ID of the subset (integer)}

\item{i18n}{Translator object from shiny.i18n library}

\item{ns}{Shiny namespace}
}
\description{
Add patients to a subset
}
\examples{
\dontrun{
patients <- c(123, 456, 789)
add_patients_to_subset(output = output, r = r, m = m, patients = patients, subset_id = 3, i18n = i18n, ns = ns)
}
}
