% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import_data.R
\name{import_dataset}
\alias{import_dataset}
\title{Import a dataset}
\usage{
import_dataset(
  omop_version = "5.4",
  data_folder = character(),
  con,
  tables_to_load = character()
)
}
\arguments{
\item{omop_version}{The OMOP version of the imported data. Accepted values are `"5.3"` and `"5.4"`. Defaults to `"5.4"`.}

\item{data_folder}{The folder containing the data. Must be a character string.}

\item{con}{A `DBI::dbConnect` object representing the database connection, required if `data_source` is `"db"`.}

\item{tables_to_load}{A character vector specifying which OMOP tables to load}
}
\description{
Imports a dataset into the application using the OMOP Common Data Model.
}
\details{
...
}
