\name{linkR-package}
\alias{linkR-package}

\alias{angleOnCircleFromPoint}
\alias{applySolveChain}
\alias{applyTransformationsChain}
\alias{avec}
\alias{avectors}
\alias{centroidSize}
\alias{circlePoint}
\alias{combineLinkages}
\alias{copyTransformation}
\alias{CSToEA}
\alias{defineCircle}
\alias{distPointToLine}
\alias{distPointToPlane}
\alias{distPointToPoint}
\alias{intersectCircles}
\alias{intersectSphereLine}
\alias{linkR_data}
\alias{linkR_examples}
\alias{minAngle}
\alias{pointNormalOnLine}
\alias{pointOnPlaneFromPoints}
\alias{pointPlaneProj}
\alias{proj3DTo2D}
\alias{reverseLinkage}
\alias{rotateBody}
\alias{rotationMatrixToEP}
\alias{setCoordinateAxes}
\alias{tMatrixDC}
\alias{tMatrixEP}
\alias{uvector}
\alias{vectorsToEP}
\alias{vorthogonal}
\alias{connJointSeq}
\alias{intersectCirclePlane}
\alias{linkageKinematics}
\alias{pointsInPlaneFromPoints}
\alias{solveKinematicChain}
\alias{sourcePartial_linkR}
\alias{writeLinkageLayers}

\docType{package}
\title{3D lever and linkage mechanism modeling}
\description{The linkR package for R provides a toolkit for creating models of 2D and 3D linkage networks, predicting motion, force and torque outputs and calculating linkage mechanical properties such as motion, force and torque transmission ratios. Users in the fields of engineering and biomechanics will find linkR particularly useful. linkR works with the R package \href{https://CRAN.R-project.org/package=svgViewR}{svgViewR}, enabling users to create 3D interactive animations of linkage models. Please see \href{https://aaronolsen.github.io/software/linkr.html}{linkR Resources} for a tutorial on how to use the linkR package and the \href{https://aaronolsen.github.io/software/linkr/examples.html}{linkR example gallery} for several examples (code and animations) of different linkages that can be created using linkR. For instructions on how to view svgViewR animations, please see \href{https://aaronolsen.github.io/software/svgviewr/interactive.html}{svgViewR interactive commands}.
}
\details{
\tabular{ll}{
Package: \tab linkR\cr
Type: \tab Package\cr
Version: \tab 1.1.1\cr
Date: \tab 2016-10-01\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Aaron Olsen
Maintainer: Aaron Olsen <aarolsen@gmail.com>
}